Create controler for task details
