+++ app/views/tasks.py
@@ -1,16 +1,11 @@
 import json
 
-from flask_restful import Resource, abort
+from flask_restful import Resource
 from flask import request
 
-from app.controllers.tasks import get_tasks_json, handle_post_task, task_schema
+from app.controllers.tasks import get_tasks_json, handle_post_task, get_task_details
 from app.models.tasks import Task
 
-def abort_if_task_not_exists(task_id):
-    task = Task.query.filter_by(id=task_id).first()
-    if not task:
-        return abort(404, message="Task with id {} does not exist.".format(task_id))
-    return task
 
 class Tasks(Resource):
     def get(self):
@@ -24,5 +19,5 @@ class Tasks(Resource):
 
 class Task_Details(Resource):
     def get(self, task_id):
-        task = abort_if_task_not_exists(task_id)
-        return task_schema.dump(task)
+        result = get_task_details(task_id)
+        return result
