+++ tests/controllers/test_tasks.py
@@ -1,0 +1,14 @@
+def test_get_tasks_json(mocker):
+    pass
+
+def test_handle_post_task(mocker):
+    pass
+
+def test_get_task_details(mocker):
+    pass
+
+def test_delete_task(mocker):
+    pass
+
+def test_update_task(mocker):
+    pass
